test(config): isolate path tests
