+++ crates/souprune/src/config.rs
@@ -456,48 +456,6 @@ pub fn resolve_path(relative_path: &str) -> Option<PathBuf> {
     None
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn desktop_projects_base_path_stays_relative() {
-        assert_eq!(get_projects_base_path(), PathBuf::from("projects"));
-    }
-
-    #[test]
-    fn android_private_base_path_points_to_app_storage() {
-        // SAFETY: These tests do not spawn threads that read this process
-        // environment while the variable is being changed.
-        unsafe {
-            std::env::remove_var("SOUPRUNE_PRIVATE_ROOT");
-        }
-        assert_eq!(
-            android_private_base_path(),
-            PathBuf::from("/data/user/0/com.bliaik.souprune/files/SoupRune")
-        );
-        assert_eq!(
-            android_private_base_path().join("projects"),
-            PathBuf::from("/data/user/0/com.bliaik.souprune/files/SoupRune/projects")
-        );
-    }
-
-    #[test]
-    fn android_private_base_path_can_come_from_environment() {
-        let root = PathBuf::from("/tmp/souprune-private-root-test");
-        // SAFETY: These tests do not spawn threads that read this process
-        // environment while the variable is being changed.
-        unsafe {
-            std::env::set_var("SOUPRUNE_PRIVATE_ROOT", &root);
-        }
-        assert_eq!(android_private_base_path(), root);
-        // SAFETY: See safety note above.
-        unsafe {
-            std::env::remove_var("SOUPRUNE_PRIVATE_ROOT");
-        }
-    }
-}
-
 /// Resource paths configuration from mod.toml [resources] section.
 ///
 /// mod.toml 中 [resources] 节的资源路径配置。
@@ -825,3 +783,6 @@ fn default_config() -> SoupruneConfig {
         resolved_dependencies: Vec::new(),
     }
 }
+
+#[cfg(test)]
+mod tests;
